Added credentials controller and method to get all credentials
